renamed 'publish_widget.py' to 'publish_frame.py'
